fix: simplify folders hierarchy
