+++ src/main/java/io/pakland/mdas/githubstats/application/FetchReviewsFromPullRequest.java
@@ -1,37 +1,35 @@
-package io.pakland.mdas.githubstats.application.external;
+package io.pakland.mdas.githubstats.application;
 
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.domain.entity.Comment;
-import io.pakland.mdas.githubstats.domain.entity.DateRange;
-import io.pakland.mdas.githubstats.domain.entity.PullRequest;
-import io.pakland.mdas.githubstats.domain.repository.CommentExternalRepository;
+import io.pakland.mdas.githubstats.domain.DateRange;
+import io.pakland.mdas.githubstats.domain.PullRequest;
+import io.pakland.mdas.githubstats.domain.Review;
+import io.pakland.mdas.githubstats.domain.repository.ReviewExternalRepository;
 
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
 
-public class FetchCommentsFromPullRequest {
+public class FetchReviewsFromPullRequest {
+    private final ReviewExternalRepository reviewExternalRepository;
 
-    private final CommentExternalRepository commentExternalRepository;
-
-    public FetchCommentsFromPullRequest(CommentExternalRepository commentExternalRepository) {
-        this.commentExternalRepository = commentExternalRepository;
+    public FetchReviewsFromPullRequest(ReviewExternalRepository reviewExternalRepository) {
+        this.reviewExternalRepository = reviewExternalRepository;
     }
 
-    public List<Comment> execute(PullRequest pullRequest, DateRange range) throws HttpException {
+    public List<Review> execute(PullRequest pullRequest, DateRange range)
+        throws HttpException {
         int page = 1, responseResults = 0;
-        List<Comment> commentList = new ArrayList<>();
+        List<Review> reviewList = new ArrayList<>();
 
         do {
-            List<Comment> apiResults = this.commentExternalRepository
-                .fetchCommentsFromPullRequestByPage(pullRequest, page);
-
+            List<Review> apiResults = this.reviewExternalRepository.fetchReviewsFromPullRequestByPage(pullRequest, page);
             if (apiResults.size() == 0) {
                 break;
             }
 
-            Instant first = apiResults.get(0).getCreatedAt().toInstant(),
-                last = apiResults.get(apiResults.size() - 1).getCreatedAt().toInstant();
+            Instant first = apiResults.get(0).getSubmittedAt().toInstant(),
+                last = apiResults.get(apiResults.size() - 1).getSubmittedAt().toInstant();
 
             // If first is before the start of the range, then we should not keep fetching the
             // data since all will be outside the range
@@ -48,22 +46,22 @@ public class FetchCommentsFromPullRequest {
             }
 
             if (range.isBetweenRange(first) && range.isBetweenRange(last)) {
-                commentList.addAll(apiResults);
+                reviewList.addAll(apiResults);
             } else {
                 // However, if first one is after the range but the last isn't, we have to filter the
                 // commits that are within the range.
                 // We have to do the same thing if the last is not in range but the first is
                 // Also if the first is after but the last is previous, there may be commits in between
                 // that are within the range
-                commentList.addAll(apiResults.parallelStream()
-                    .filter(commit -> range.isBetweenRange(commit.getCreatedAt().toInstant()))
+                reviewList.addAll(apiResults.parallelStream()
+                    .filter(review -> range.isBetweenRange(review.getSubmittedAt().toInstant()))
                     .toList());
             }
 
             responseResults = apiResults.size();
             page++;
-        } while (responseResults > 0);
+        } while (responseResults != 100);
 
-        return commentList;
+        return reviewList;
     }
 }
