+++ src/main/java/io/pakland/mdas/githubstats/application/FetchTeamsFromOrganization.java
@@ -1,8 +1,8 @@
-package io.pakland.mdas.githubstats.application.external;
+package io.pakland.mdas.githubstats.application;
 
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.domain.entity.Organization;
-import io.pakland.mdas.githubstats.domain.entity.Team;
+import io.pakland.mdas.githubstats.domain.Organization;
+import io.pakland.mdas.githubstats.domain.Team;
 import io.pakland.mdas.githubstats.domain.repository.TeamExternalRepository;
 
 import java.util.List;
