+++ src/main/java/io/pakland/mdas/githubstats/application/FetchUsersFromTeam.java
@@ -1,9 +1,10 @@
-package io.pakland.mdas.githubstats.application.external;
+package io.pakland.mdas.githubstats.application;
 
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.domain.entity.Team;
-import io.pakland.mdas.githubstats.domain.entity.User;
+import io.pakland.mdas.githubstats.domain.Team;
+import io.pakland.mdas.githubstats.domain.User;
 import io.pakland.mdas.githubstats.domain.repository.UserExternalRepository;
+
 import java.util.List;
 
 public class FetchUsersFromTeam {
