+++ src/main/java/io/pakland/mdas/githubstats/application/dto/PullRequestStateDTO.java
@@ -1,6 +1,6 @@
 package io.pakland.mdas.githubstats.application.dto;
 
-import io.pakland.mdas.githubstats.domain.enums.PullRequestState;
+import io.pakland.mdas.githubstats.domain.PullRequestState;
 
 public interface PullRequestStateDTO {
     PullRequestState getValue();
