+++ src/main/java/io/pakland/mdas/githubstats/application/mappers/RepositoryMapper.java
@@ -1,7 +1,7 @@
 package io.pakland.mdas.githubstats.application.mappers;
 
 import io.pakland.mdas.githubstats.application.dto.RepositoryDTO;
-import io.pakland.mdas.githubstats.domain.entity.Repository;
+import io.pakland.mdas.githubstats.domain.Repository;
 
 public class RepositoryMapper {
