+++ src/main/java/io/pakland/mdas/githubstats/application/mappers/ReviewMapper.java
@@ -1,7 +1,7 @@
 package io.pakland.mdas.githubstats.application.mappers;
 
 import io.pakland.mdas.githubstats.application.dto.ReviewDTO;
-import io.pakland.mdas.githubstats.domain.entity.Review;
+import io.pakland.mdas.githubstats.domain.Review;
 
 public class ReviewMapper {
