+++ src/main/java/io/pakland/mdas/githubstats/application/mappers/TeamMapper.java
@@ -1,7 +1,7 @@
 package io.pakland.mdas.githubstats.application.mappers;
 
 import io.pakland.mdas.githubstats.application.dto.TeamDTO;
-import io.pakland.mdas.githubstats.domain.entity.Team;
+import io.pakland.mdas.githubstats.domain.Team;
 
 public class TeamMapper {
     public static Team dtoToEntity(TeamDTO dto) {
