+++ src/main/java/io/pakland/mdas/githubstats/application/mappers/UserMapper.java
@@ -1,8 +1,7 @@
 package io.pakland.mdas.githubstats.application.mappers;
 
 import io.pakland.mdas.githubstats.application.dto.UserDTO;
-import io.pakland.mdas.githubstats.domain.entity.User;
-import java.util.ArrayList;
+import io.pakland.mdas.githubstats.domain.User;
 
 public class UserMapper {
     public static User dtoToEntity(UserDTO dto) {
