+++ src/main/java/io/pakland/mdas/githubstats/domain/Comment.java
@@ -1,7 +1,11 @@
-package io.pakland.mdas.githubstats.domain.entity;
+package io.pakland.mdas.githubstats.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.Date;
-import lombok.*;
 
 @Data
 @NoArgsConstructor
