+++ src/main/java/io/pakland/mdas/githubstats/domain/DateRange.java
@@ -1,7 +1,11 @@
-package io.pakland.mdas.githubstats.domain.entity;
+package io.pakland.mdas.githubstats.domain;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 import java.time.Instant;
-import lombok.*;
 
 @Builder
 @NoArgsConstructor
