+++ src/main/java/io/pakland/mdas/githubstats/domain/EntityType.java
@@ -1,4 +1,4 @@
-package io.pakland.mdas.githubstats.domain.enums;
+package io.pakland.mdas.githubstats.domain;
 
 public enum EntityType {
     ORGANIZATION,
