+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/CommentExternalRepository.java
@@ -1,8 +1,8 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.domain.entity.Comment;
-import io.pakland.mdas.githubstats.domain.entity.PullRequest;
+import io.pakland.mdas.githubstats.domain.Comment;
+import io.pakland.mdas.githubstats.domain.PullRequest;
 
 import java.util.List;
