+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/RepositoryExternalRepository.java
@@ -1,8 +1,8 @@
 package io.pakland.mdas.githubstats.domain.repository;
 
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
-import io.pakland.mdas.githubstats.domain.entity.Repository;
-import io.pakland.mdas.githubstats.domain.entity.Team;
+import io.pakland.mdas.githubstats.domain.Repository;
+import io.pakland.mdas.githubstats.domain.Team;
 
 import java.util.List;
