+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/github/model/GitHubPullRequestStateDTO.java
@@ -2,7 +2,7 @@ package io.pakland.mdas.githubstats.infrastructure.github.model;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.pakland.mdas.githubstats.application.dto.PullRequestStateDTO;
-import io.pakland.mdas.githubstats.domain.enums.PullRequestState;
+import io.pakland.mdas.githubstats.domain.PullRequestState;
 import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
