+++ src/main/java/io/pakland/mdas/githubstats/infrastructure/github/repository/OrganizationGitHubRepository.java
@@ -2,13 +2,14 @@ package io.pakland.mdas.githubstats.infrastructure.github.repository;
 
 import io.pakland.mdas.githubstats.application.exceptions.HttpException;
 import io.pakland.mdas.githubstats.application.mappers.OrganizationMapper;
-import io.pakland.mdas.githubstats.domain.entity.Organization;
+import io.pakland.mdas.githubstats.domain.Organization;
 import io.pakland.mdas.githubstats.domain.repository.OrganizationExternalRepository;
 import io.pakland.mdas.githubstats.infrastructure.github.model.GitHubOrganizationDTO;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.web.reactive.function.client.WebClientResponseException;
+
+import java.util.List;
 
 public class OrganizationGitHubRepository implements OrganizationExternalRepository {
